refactor(cli): group helper ssh commands into ssh sub-package
