+++ cmd/internal/helper/ssh/ssh_git_clone.go
@@ -1,4 +1,4 @@
-package helper
+package ssh
 
 import (
 	"context"
